Move registration React app to its own controller
